Fixed typo english.json / update polish.json (#82)
* Finished spanish translation and fixed typos

There are possible minor fixed that could be made for better understanding that are nitpicking in a way, but if there is a certain demand or time available i will fix it in another commit.

* Updated spanish.json [line 434]

* spelling check finished

Grammar corrected, added accents where needed, Correct message expression related to english added/modified, last thing needed to check is if someone translated a command, that could cause confusion, first impression is that this case it's not present at the moment, when confirmed, i'll send a message on discord, otherwise i will make a commit.

* Update spanish.json

* Removing useless space

* Update spanish.json

* translation in portuguese finished

* Update russian.json

* Update english.json

* added arabic.json file

* Useless if it's fully in english

* fixed typo on english.json[line 325]



*chose instead of choosed

Certain words we tend to add "ed" to sound past tense. But not all. "Choosed" is grammatically wrong English(In this case choosed  as a word doesn't exists.

* fixed typo on english.json[line 139]

Changed states with status

* Fixed typo english.json[line 325]

*chose instead of choosed

Certain words we tend to add "ed" to sound past tense. But not all. "Choosed" is grammatically wrong English(In this case choosed  as a word doesn't exists.