Changes on music-autoleave-never
I'm totally made a mistake on music-autoleave-never. Actually, the word for never is "tidak pernah" and if it based on the context is "Tidak akan meninggakan" or similar to that. But, in the code, I wrote never as "jangan pernah" that also have the same meaning but it would make people ambiguous.